smartclide-wizard/requirements.txt
setuptools
cookiecutter
argparse
pyyaml